% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ipw_viz_bar}
\alias{ipw_viz_bar}
\title{Creating bar charts to check if the IPW weighting actually removed the relationship between treatment and the intended variable}
\usage{
ipw_viz_bar(data, x, fill, weights)
}
\arguments{
\item{data}{the data used in the IPW analysis}

\item{x}{the x axis for the boxplot (the variable we want to no longer have a relationship with treatment)}

\item{fill}{the fill for the boxplot (typically the treatment)}

\item{weights}{optional parameter to include the weights}
}
\description{
Creating bar charts to check if the IPW weighting actually removed the relationship between treatment and the intended variable
}
\examples{
ipw_viz_bar(sim_data, Z, A)
ipw_viz_bar(sim_data, Z, A, ipw_complex)
}
