% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_x_in_collider}
\alias{set_x_in_collider}
\title{Simulates X values in a collider. This function is used within the simple_collider_simulation(n, x_numeric, y_numeric, z_numeric) function.}
\usage{
set_x_in_collider(n, x_numeric)
}
\arguments{
\item{n}{The number of observations of X for a single simulation}

\item{x_numeric}{TRUE for numeric X and FALSE for binary X}
}
\value{
Simulated X values
}
\description{
Simulates X values in a collider. This function is used within the simple_collider_simulation(n, x_numeric, y_numeric, z_numeric) function.
}
