% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_z_in_collider}
\alias{set_z_in_collider}
\title{Simulates Z values in a collider. This function is used within the simple_collider_simulation(n, x_numeric, y_numeric, z_numeric) function.}
\usage{
set_z_in_collider(n, z_numeric)
}
\arguments{
\item{n}{The number of observations of Z for a single simulation}

\item{z_numeric}{TRUE for numeric Z and FALSE for binary Z}
}
\value{
Simulated Z values
}
\description{
Simulates Z values in a collider. This function is used within the simple_collider_simulation(n, x_numeric, y_numeric, z_numeric) function.
}
