% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_from_str}
\alias{sim_from_str}
\title{Simulates data based on the structural equations provided by the user}
\usage{
sim_from_str(N, string)
}
\arguments{
\item{N}{the number of observations for the data}

\item{string}{a string containing the structural equation or type of distribution for each variable, as well as the type of data they want in parentheses, comma separated for each variable.}
}
\value{
A data frame with data generated as specified by the structural equations provided.
}
\description{
This function simulates causal data from a string given by the user. There are a few important things that must be true of the string. The format should be the name of the variable ~ the equation of the variable (binomial, normal, or an equation like x+5) and then the type of output in parenthesis (quant, binary, linear). Each variable should be comma separated, and the variables should be in order with the furthest ancestor first and the furthest descendant last.
}
\examples{
# example code
sim_data <- sim_from_str(100, "z ~ normal (quant), x ~ binomial (bin), y ~ x + z (linear)")
sim_from_str(1000, "a ~ binomial (bin), b ~ 5a + 30 (bin), c~ a+b+20 (quant), d ~ a + b + c (lin)")
}
