% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simple_fork_simulation}
\alias{simple_fork_simulation}
\title{Simulates the dependency structure of a basic fork, with 3 variables X, Y, Z, where X ← Y → Z}
\usage{
simple_fork_simulation(n, x_numeric, y_numeric, z_numeric)
}
\arguments{
\item{n}{The number of observations of X, Y, Z for a single simulation}

\item{x_numeric}{TRUE for numeric X and FALSE for binary X}

\item{y_numeric}{TRUE for numeric Y and FALSE for binary Y}

\item{z_numeric}{TRUE for numeric Z and FALSE for binary Z}
}
\value{
A data frame of the 4 tests performed (Check if X and Y are marginally dependent, Check if Z and Y are marginally dependent, Check if X and Z are marginally dependent, Check if X and Z are conditionally independent given Y), the model p-values corresponding to the tests, and the interpretation of the result.
}
\description{
This function takes the number of observations of X, Y, Z for a single simulation. It also requires the user to specify whether they want X, Y, or Z to be numeric or binary
}
\examples{
# example code
# Results in a fork simulation with 10000 simulated observations of X, Y, Z, with X and Y being numeric and Z being binary.
set.seed(451)
simple_fork_simulation(10000, x_numeric = TRUE, y_numeric = TRUE, z_numeric = FALSE)
}
