% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_many}
\alias{simulate_many}
\title{Function simulator: Simulates a key structure simulation function of choice for n number of times}
\usage{
simulate_many(func, n_rep_time)
}
\arguments{
\item{func}{A function out of the 3 configured functions: simple_chain_simulation(n, x_numeric, y_numeric, z_numeric), simple_fork_simulation(n, x_numeric, y_numeric, z_numeric), simple_collider_simulation(n, x_numeric, y_numeric, z_numeric)}

\item{n_rep_time}{The number of times a function is replicated}
}
\value{
A data frame of the 4 tests performed (Check if X and Y are marginally dependent, Check if Z and Y are marginally dependent, Check if X and Z are marginally dependent, Check if X and Z are conditionally independent given Y), the percentage of significant p-values corresponding to the tests, and the interpretation of the result.
}
\description{
This function replicates n number of times a chosen function out of 3 functions: simple_chain_simulation(n, x_numeric, y_numeric, z_numeric), simple_fork_simulation(n, x_numeric, y_numeric, z_numeric), simple_collider_simulation(n, x_numeric, y_numeric, z_numeric).
}
\examples{
# example code
# A replication of the simple_fork_simulation(10000, x_numeric = TRUE, y_numeric = FALSE, z_numeric = TRUE) for 100 times
set.seed(451)
simulate_many(simple_fork_simulation(10000, x_numeric = TRUE, y_numeric = FALSE, z_numeric = TRUE), n_rep_time = 100)
}
