% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weighting}
\alias{weighting}
\title{Creating the weights used for inverse probability weighting or propensity score weighting}
\usage{
weighting(data, model, output)
}
\arguments{
\item{data}{the data used to create the weights}

\item{model}{the model which we want to apply the weights to}

\item{output}{what we are trying to model (for IPW it should be the treatment {A}; for propensity scoring it should be our outcome variable {Y})}
}
\description{
Creating the weights used for inverse probability weighting or propensity score weighting
}
\examples{
weighting(sim_data, ps_mod_simple, A)
weighting(tenure, ps_mod, tenure$gncs)
}
